feat: waitForCoreChainLockSyncFactory
