+++ lib/abci/handlers/initChainHandlerFactory.js
@@ -3,7 +3,6 @@ const {
     ResponseInitChain,
   },
 } = require('abci/types');
-const { ChainLock } = require('@dashevo/dashcore-lib');
 const waitForEvent = require('../../util/waitForEvent');
 /**
  * Init Chain ABCI handler
@@ -31,30 +30,6 @@ function initChainHandlerFactory(
    */
   async function initChainHandler() {
     logger.info('Init chain');
-
-    // We need to retrieve latest chainlock from our fully synced DashCore instance
-    const rpcBestChainLockResponse = await coreRpcClient.getBestChainLock();
-
-    const chainlock = new ChainLock(rpcBestChainLockResponse);
-    latestCoreChainLock.update(chainlock);
-
-    const zmqSocket = coreZMQClient.connect();
-    zmqSocket.subscribe(coreZMQClient.ROOMS.rawchainlock);
-    logger.info('Subscribe to rawchainlock ZMQ room');
-
-    // listens to ChainLock zmq zmqpubrawchainlocksig event and updates the latestCoreChainLock.
-    // If we lost connection with Core (heartbeats with retry policy) we need to exit with fatal
-    // In case of regtest fallback we need to listen for new blocks using hashblock zmq event
-    zmqSocket.on(coreZMQClient.ROOMS.rawchainlock, (rawChainLockMessage) => {
-      const socketChainLock = new ChainLock(rawChainLockMessage);
-      logger.info(`Received new chainlock ${socketChainLock.toString()}`);
-      latestCoreChainLock.update(socketChainLock);
-    });
-
-    if (!latestCoreChainLock.chainLock) {
-      // We need to wait for new chainlock from ZMQ socket
-      await waitForEvent(zmqSocket, coreZMQClient.ROOMS.rawchainlock);
-    }
 
     return new ResponseInitChain({
       latestCoreChainLock,
