+++ lib/core/LatestCoreChainLock.js
@@ -4,8 +4,8 @@ class LatestCoreChainLock {
    * @param {string} [chainLock]
    * @param {LatestCoreChainLock}
    */
-  constructor(chainLock) {
-    this.chainLock = chainLock || null;
+  constructor(chainLock = undefined) {
+    this.chainLock = chainLock || undefined;
   }
 
   /**
