+++ lib/createDIContainer.js
@@ -14,6 +14,7 @@ const level = require('level-rocksdb');
 const LRUCache = require('lru-cache');
 
 const RpcClient = require('@dashevo/dashd-rpc/promise');
+const ZMQClient = require('@dashevo/dashd-zmq');
 
 const DashPlatformProtocol = require('@dashevo/dpp');
 
@@ -24,7 +25,6 @@ const pino = require('pino');
 const sanitizeUrl = require('./util/sanitizeUrl');
 
 const LatestCoreChainLock = require('./core/LatestCoreChainLock');
-const ZMQClient = require('./core/ZMQClient');
 const IdentityLevelDBRepository = require('./identity/IdentityLevelDBRepository');
 const PublicKeyIdentityIdMapLevelDBRepository = require(
   './identity/PublicKeyIdentityIdMapLevelDBRepository',
@@ -79,6 +79,7 @@ const endBlockHandlerFactory = require('./abci/handlers/endBlockHandlerFactory')
 const queryHandlerFactory = require('./abci/handlers/queryHandlerFactory');
 
 const waitForCoreSyncFactory = require('./core/waitForCoreSyncFactory');
+const waitForCoreChainLockSyncFactory = require('./core/waitForCoreChainLockSyncFactory');
 
 /**
  *
@@ -159,6 +160,13 @@ async function createDIContainer(options) {
   });
 
   /**
+   * Register Core related
+   */
+  container.register({
+    latestCoreChainLock: asClass(LatestCoreChainLock).singleton(),
+  });
+
+  /**
    * Register common services
    */
   container.register({
@@ -200,13 +208,6 @@ async function createDIContainer(options) {
         user: coreJsonRpcUsername,
         pass: coreJsonRpcPassword,
       }))).singleton(),
-  });
-
-  /**
-   * Register Core related
-   */
-  container.register({
-    latestCoreChainLock: asClass(LatestCoreChainLock).singleton(),
   });
 
   /**
@@ -408,7 +409,9 @@ async function createDIContainer(options) {
   container.register({
     waitForCoreSync: asFunction(waitForCoreSyncFactory).singleton(),
   });
-
+  container.register({
+    waitForCoreChainLockSync: asFunction(waitForCoreChainLockSyncFactory).singleton(),
+  });
   container.register({
     waitReplicaSetInitialize: asFunction(waitReplicaSetInitializeFactory).singleton(),
   });
